+++ client/components/index.js
@@ -10,3 +10,4 @@ export {default as FriendList} from './friendList'
 export {default as MyProfile} from './myprofile'
 export {default as Main} from './mainComponent/main'
 export {default as AddFriend} from './addFriend'
+export {default as UpdateForm} from './updateUserForm'
