+++ client/store/index.js
@@ -2,18 +2,14 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import {userFriendsReducer} from './reducers'
-import convo from './convo'
-import message from './message'
-import friend from './friend'
+import {userFriends, user, convo, message, friend} from './reducers'
 
 const reducer = combineReducers({
   user,
   convo,
   message,
   friend,
-  userFriends: userFriendsReducer
+  userFriends
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
@@ -21,4 +17,4 @@ const middleware = composeWithDevTools(
 const store = createStore(reducer, middleware)
 
 export default store
-export * from './user'
+export * from './reducers/user'
